Add flexibility is seconds and milliseconds format
