+++ tests/data/2.preprocessed_data/PPG/PPG_meta.json
@@ -6,24 +6,25 @@
     "metadata_version": "0.1",
     "start_iso8601": "2021-06-27T16:52:20Z",
     "end_iso8601": "2021-06-27T17:03:05Z",
-    "scale_factors": [
-        1
-    ],
     "rows": 19338,
     "endianness": "little",
-    "data_type": "float",
     "bits": 64,
+    "data_type": "float",
     "sensors": [
         {
+            "scale_factors": [
+                1
+            ],
             "file_name": "PPG_time.bin",
             "channels": [
                 "time"
             ],
             "units": [
-                "absolute_ms"
+                "absolute_s"
             ]
         },
         {
+            "scale_factors": [],
             "file_name": "PPG_values.bin",
             "channels": [
                 "green"
