+++ tests/data/2.preprocessed_data/PPG/accelerometer_meta.json
@@ -20,7 +20,7 @@
                 "time"
             ],
             "units": [
-                "absolute_ms"
+                "absolute_s"
             ]
         },
         {
@@ -35,14 +35,7 @@
                 "m/s^2",
                 "m/s^2"
             ],
-            "scale_factors": [
-                0.00469378,
-                0.00469378,
-                0.00469378,
-                0.06097561,
-                0.06097561,
-                0.06097561
-            ]
+            "scale_factors": []
         }
     ]
 }
