+++ tests/data/2.preprocessed_data/imu/accelerometer_meta.json
@@ -20,7 +20,7 @@
                 "time"
             ],
             "units": [
-                "time_relative_ms"
+                "relative_s"
             ]
         },
         {
